+++ www/whatsnew.md
@@ -20,6 +20,7 @@ Thanks to a great contribution from @doombubbles, new custom cards and a signifi
  - Elortha no Shanda now does not cause infinite deathrattle loops.
  - Dinosaur Chops is now uncollectible.
  - Thunderfury now is immune and does not reduce durability while you are overloaded.
+ - Shudderwock now triggers Hagatha the Witch and other hero battlecries.
 
 ###### Changes Since 0.6.5-2.0.11 (Tuesday, November 27th, 2018)